+++ database/migrations/2023_01_16_110253_create_seguimientos_table.php
@@ -15,12 +15,12 @@ class CreateSeguimientosTable extends Migration
     {
         Schema::create('seguimientos', function (Blueprint $table) {
             $table->id();
-            $table->integer('idProfesor');
+            $table->unsignedBigInteger('idProfesor');
             $table->foreign("idProfesor")->references('id')->on('profesors')->onUpdate("cascade")->onDelete("cascade");
-            $table->integer('idAlumno');
+            $table->unsignedBigInteger('idAlumno');
             $table->foreign("idAlumno")->references('id')->on('alumnos')->onUpdate("cascade")->onDelete("cascade");
             $table->dateTime('fechaHora');
-            $table->integer('idAsignatura');
+            $table->unsignedBigInteger('idAsignatura');
             $table->foreign("idAsignatura")->references('id')->on('asignaturas')->onUpdate("cascade")->onDelete("cascade");
             $table->string('comentario');
             $table->primary(["idProfesor", "idAlumno", "fechaHora", "idAsignatura"]);
